requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2 
django-import-export>=3.3.4
qrcode[pil]>=7.4.2
mysqlclient>=2.2.0
